+++ tests/utils/test_concurrent.py
@@ -4,7 +4,7 @@ from concurrent.futures import TimeoutError
 
 import pytest
 
-from arroyo.utils.concurrent import Synchronized, execute
+from arroyo.utils.concurrent import execute
 
 
 def test_execute() -> None:
@@ -15,21 +15,3 @@ def test_execute() -> None:
 
     with pytest.raises(TimeoutError):
         assert execute(lambda: time.sleep(10), daemon=True).result(timeout=0)
-
-
-def test_synchronized() -> None:
-    value = object()
-    wrapper = Synchronized(value)
-
-    with wrapper.get() as wrapped:
-        assert wrapped is value
-
-    wrapper.set(object())
-
-    with wrapper.get() as wrapped:
-        assert wrapped is not value
-
-    wrapper.set(value)
-
-    with wrapper.get() as wrapped:
-        assert wrapped is value
